Fix rm calls to properly handle all cases
In previous iterations of these scripts, it seems that the archives used
to be called `evm_full_archive*` and `evm_full_archived*` (with a 'd'),
even though neither of the two scripts generate these file names.  The
former was being deleted in both scripts, and the latter was being
deleted in the `archive` script.

On top of that, `evm_full_archive*` (without a 'd') was being generated
via the archive script, but not being cleared out by the `current` or
`archive` scripts, and only the `current` script cleaned up any
`current` tar files.

This changes both of the scripts `rm -rf` calls to match each other so
they don't miss anything in the cleanup.