DRY up collection scripts
Both the `archive` and `current` scripts share 95% of the same code, so
this takes the majority of that code, and puts it in a combined script,
and adds some variables to assist in the parts that change between the
two.  Those can be set within the scripts themselves, or even added to
the main script instead.

A wrapper script is current required to handle the `exclude_files` file,
since it needs to be read from the directory that the script is coming
from.  This will most likely be changed in the future to make require
less boiler plate code in the helper scripts.