Clean/DRY up tar command
The `tar` command that is executed with this script does not change when
we are on a database appliance or not.  This moves the files that are
being exported to a variable that can then be added to if we are on a
database.

Also, this change adds a bit of documentation to the script as to what
files are being exported.